% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plots.R
\name{combine_auto_plots}
\alias{combine_auto_plots}
\title{Combine All Automatic Plots}
\usage{
combine_auto_plots(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
A combined patchwork plot
}
\description{
Combine All Automatic Plots
}
