% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_detect.R
\name{detect_funding_themes_qdap}
\alias{detect_funding_themes_qdap}
\title{Detect Funding Themes Using qdap}
\usage{
detect_funding_themes_qdap(data, purpose_column = "purpose")
}
\arguments{
\item{data}{A dataset containing a purpose text column}

\item{purpose_column}{Name of the Purpose column (string)}
}
\value{
Data with added `theme` column
}
\description{
Uses rule-based keyword dictionaries to classify grant purposes
into interpretable usual NGO themes.
}
