% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plots.R
\name{export_auto_plots_pdf}
\alias{export_auto_plots_pdf}
\title{Print all of the plots as a PDF}
\usage{
export_auto_plots_pdf(data, filename = "auto_plots.pdf")
}
\arguments{
\item{data}{A dataset}

\item{filename}{Name of the output PDF file}
}
\description{
Print all of the plots as a PDF
}
