% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_detect.R
\name{ggplot_funding_themes}
\alias{ggplot_funding_themes}
\title{Visualize the Detected Funding Themes
ggplot bar chart of funding themes}
\usage{
ggplot_funding_themes(data, theme_column = "theme_qdap")
}
\arguments{
\item{data}{A dataset with detected themes}

\item{theme_column}{Name of the theme column (string)}
}
\value{
A ggplot object
}
\description{
Visualize the Detected Funding Themes
ggplot bar chart of funding themes
}
