% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plots.R
\name{summarize_amount_avg_year}
\alias{summarize_amount_avg_year}
\title{Summarize Average Amount by Year}
\usage{
summarize_amount_avg_year(data, x = NULL, y = NULL)
}
\arguments{
\item{data}{A dataset}

\item{x}{Name of year column (string)}

\item{y}{Name of amount column (string)}
}
\value{
A ggplot object
}
\description{
Summarize Average Amount by Year
}
