% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plots.R
\name{summarize_location_count}
\alias{summarize_location_count}
\title{Summarize Location Count}
\usage{
summarize_location_count(data, x = NULL)
}
\arguments{
\item{data}{A dataset}

\item{x}{Name of location column (string)}
}
\value{
A ggplot object
}
\description{
Summarize Location Count
}
