% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plots.R
\name{summarize_orgsize_amount}
\alias{summarize_orgsize_amount}
\title{Organization Size vs Amount}
\usage{
summarize_orgsize_amount(data, x = NULL, y = NULL)
}
\arguments{
\item{data}{A dataset}

\item{x}{Name of org size column (string)}

\item{y}{Name of amount column (string)}
}
\value{
A ggplot object
}
\description{
Organization Size vs Amount
}
