% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_plots.R
\name{time_series_amount_year}
\alias{time_series_amount_year}
\title{Time Series of Amount Over Year}
\usage{
time_series_amount_year(data, x = "year", y = "amount")
}
\arguments{
\item{data}{A dataset}

\item{x}{Name of year column (string)}

\item{y}{Name of amount column (string)}
}
\value{
A ggplot object
}
\description{
Time Series of Amount Over Year
}
