% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data.R
\name{validate_data}
\alias{validate_data}
\title{Validate the data condition in the uploaded dataset}
\usage{
validate_data(data)
}
\arguments{
\item{data}{A data frame containing the uploaded dataset.}
}
\value{
A list containing a boolean 'is_valid' and a character vector 'messages

Summary of the dataset
}
\description{
Validate the data condition in the uploaded dataset
}
